fix: correções dos filtros nos repositorios e nos services

+++ vox/src/main/java/br/edu/ifpb/pweb2/vox/repository/AssuntoRepository.java
@@ -8,7 +8,6 @@ import br.edu.ifpb.pweb2.vox.entity.Assunto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface AssuntoRepository extends JpaRepository<Assunto, Integer> {
-
+public interface AssuntoRepository extends JpaRepository<Assunto, Long> {
     
 }
